projects_2024/project1_rlhf_summaries/requirements.txt
google-cloud-aiplatform 
google-cloud-pipeline-components
kfp
gcloud
tensorboard
pandas
rouge